requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
idna==2.6
mysqlclient==1.3.13
paramiko==2.4.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.5
requests==2.18.4
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.11
sshtunnel==0.1.4
urllib3==1.22
